Extract cloud runtime dispatch and shared helpers
